+++ Abilities/AirJump.cs
@@ -1,5 +1,6 @@
-using System.IO;
+using AnimLib.Abilities;
 using OriMod.Utilities;
+using System.IO;
 using Terraria;
 using Terraria.ModLoader;
 
@@ -7,14 +8,15 @@ namespace OriMod.Abilities {
   /// <summary>
   /// Ability for jumping in the air.
   /// </summary>
-  public sealed class AirJump : Ability, ILevelable {
-    internal AirJump(AbilityManager manager) : base(manager) { }
+  public sealed class AirJump : Ability<OriAbilityManager>, ILevelable {
     public override int Id => AbilityId.AirJump;
-    public override byte Level => (this as ILevelable).Level;
-    byte ILevelable.Level { get; set; }
-    public byte MaxLevel => 4;
+    public override int Level => (this as ILevelable).Level;
+    public override bool Unlocked => Level > 0;
+    int ILevelable.Level { get; set; }
+    public int MaxLevel => 4;
 
-    internal override bool CanUse => base.CanUse && !oPlayer.IsGrounded && !oPlayer.OnWall && currentCount < MaxJumps && !player.mount.Active &&
+    public override bool CanUse => base.CanUse && !abilities.oPlayer.IsGrounded && !abilities.oPlayer.OnWall &&
+      currentCount < MaxJumps && !player.mount.Active &&
       !abilities.bash && !abilities.burrow && !abilities.climb && !abilities.chargeJump && !abilities.launch &&
       !abilities.wallChargeJump;
 
@@ -25,15 +27,15 @@ namespace OriMod.Abilities {
     internal ushort currentCount;
     private sbyte _gravityDirection;
 
-    protected override void ReadPacket(BinaryReader r) {
-      currentCooldown = r.ReadUInt16();
+    public override void ReadPacket(BinaryReader r) {
+      cooldownLeft = r.ReadInt32();
       _gravityDirection = r.ReadSByte();
       player.position = r.ReadVector2();
       player.velocity = r.ReadVector2();
     }
 
-    protected override void WritePacket(ModPacket packet) {
-      packet.Write(currentCount);
+    public override void WritePacket(ModPacket packet) {
+      packet.Write(cooldownLeft);
       packet.Write(_gravityDirection);
       packet.WriteVector2(player.position);
       packet.WriteVector2(player.velocity);
@@ -41,39 +43,39 @@ namespace OriMod.Abilities {
 
     private readonly RandomChar _rand = new RandomChar();
 
-    protected override void UpdateActive() {
-      float newVel = -JumpVelocity * ((float)(EndDuration - CurrentTime) / EndDuration) * _gravityDirection;
+    public override void UpdateActive() {
+      float newVel = -JumpVelocity * ((float)(EndDuration - stateTime) / EndDuration) * _gravityDirection;
       player.velocity.Y = newVel;
     }
 
-    internal override void Tick() {
-      if (CanUse && input.jump.JustPressed) {
+    public override void PreUpdate() {
+      if (CanUse && abilities.oPlayer.input.jump.JustPressed) {
         if (player.canJumpAgain_Blizzard || player.canJumpAgain_Cloud || player.canJumpAgain_Fart || player.canJumpAgain_Sail ||
-            player.canJumpAgain_Sandstorm || player.canCarpet || (player.rocketBoots!=0 && player.rocketTime>0) || player.mount.Active) return;
-        SetState(State.Active);
+            player.canJumpAgain_Sandstorm || player.canCarpet || (player.rocketBoots != 0 && player.rocketTime > 0) || player.mount.Active) return;
+        SetState(AbilityState.Active);
         currentCount++;
         _gravityDirection = (sbyte)player.gravDir;
 
         if (MaxJumps != 1 && currentCount == MaxJumps) {
-          oPlayer.PlaySound("Ori/TripleJump/seinTripleJumps" + _rand.NextNoRepeat(5), 0.6f);
+          abilities.oPlayer.PlaySound("Ori/TripleJump/seinTripleJumps" + _rand.NextNoRepeat(5), 0.6f);
         }
         else {
-          oPlayer.PlaySound("Ori/DoubleJump/seinDoubleJumps" + _rand.NextNoRepeat(4), 0.5f);
+          abilities.oPlayer.PlaySound("Ori/DoubleJump/seinDoubleJumps" + _rand.NextNoRepeat(4), 0.5f);
         }
         return;
       }
-      if (oPlayer.IsGrounded || abilities.bash || abilities.launch || oPlayer.OnWall) {
+      if (abilities.oPlayer.IsGrounded || abilities.bash || abilities.launch || abilities.oPlayer.OnWall) {
         currentCount = 0;
-        if (oPlayer.IsGrounded || abilities.bash || abilities.launch || abilities.climb) {
-          SetState(State.Inactive);
+        if (abilities.oPlayer.IsGrounded || abilities.bash || abilities.launch || abilities.climb) {
+          SetState(AbilityState.Inactive);
         }
       }
       if (Active) {
-        SetState(State.Ending);
+        SetState(AbilityState.Ending);
       }
       else if (Ending) {
-        if (CurrentTime > EndDuration || player.velocity.Y * player.gravDir > 0) {
-          SetState(State.Inactive);
+        if (stateTime > EndDuration || player.velocity.Y * player.gravDir > 0) {
+          SetState(AbilityState.Inactive);
         }
       }
       // Other than activation, Air Jump is deterministic and requires no additional syncing