+++ Abilities/Stomp.cs
@@ -1,8 +1,8 @@
-//using AnimLib.Abilities;
-using System;
+using AnimLib.Abilities;
 using Microsoft.Xna.Framework;
 using OriMod.Projectiles.Abilities;
 using OriMod.Utilities;
+using System;
 using Terraria;
 using Terraria.Graphics.Shaders;
 using Terraria.ID;
@@ -11,18 +11,18 @@ namespace OriMod.Abilities {
   /// <summary>
   /// Ability for an air-to-ground Area of Effect attack.
   /// </summary>
-  public sealed class Stomp : Ability, ILevelable {
-    internal Stomp(AbilityManager manager) : base(manager) { }
+  public sealed class Stomp : Ability<OriAbilityManager>, ILevelable {
     public override int Id => AbilityId.Stomp;
-    public override byte Level => (this as ILevelable).Level;
-    byte ILevelable.Level { get; set; }
-    byte ILevelable.MaxLevel => 3;
+    public override int Level => (this as ILevelable).Level;
+    int ILevelable.Level { get; set; }
+    int ILevelable.MaxLevel => 3;
+    public override bool Unlocked => Level > 0;
 
-    internal override bool CanUse => base.CanUse && !oPlayer.IsGrounded && !InUse && !player.mount.Active && player.grapCount == 0 &&
+    public override bool CanUse => base.CanUse && !abilities.oPlayer.IsGrounded && !InUse && !player.mount.Active && player.grapCount == 0 &&
       !abilities.bash && !abilities.burrow && !abilities.chargeDash && !abilities.chargeJump && !abilities.climb &&
       !abilities.dash && !abilities.glide.Active && !abilities.launch && !abilities.wallChargeJump;
-    protected override int Cooldown => Math.Min(30 + Level * 30, 600);
-    protected override Color RefreshColor => Color.Orange;
+    public override int Cooldown => Math.Min(30 + Level * 30, 600);
+    public override void OnRefreshed() => abilities.RefreshParticles(Color.Orange);
 
     private int Damage => 30 + Level * 20;
 
@@ -61,18 +61,18 @@ namespace OriMod.Abilities {
     private readonly RandomChar _randActive = new RandomChar();
     private readonly RandomChar _randEnd = new RandomChar();
 
-    protected override void UpdateStarting() {
-      if (CurrentTime == 0) {
-        oPlayer.PlaySound("Ori/Stomp/seinStompStart" + _randStart.NextNoRepeat(3), 0.8f, 0.2f);
+    public override void UpdateStarting() {
+      if (stateTime == 0) {
+        abilities.oPlayer.PlaySound("Ori/Stomp/seinStompStart" + _randStart.NextNoRepeat(3), 0.8f, 0.2f);
       }
       player.velocity.X = 0;
       player.velocity.Y *= 0.9f;
       player.gravity = -0.1f;
     }
 
-    protected override void UpdateActive() {
-      if (CurrentTime == 0) {
-        oPlayer.PlaySound("Ori/Stomp/seinStompFall" + _randActive.NextNoRepeat(3), 0.8f);
+    public override void UpdateActive() {
+      if (stateTime == 0) {
+        abilities.oPlayer.PlaySound("Ori/Stomp/seinStompFall" + _randActive.NextNoRepeat(3), 0.8f);
         NewAbilityProjectile<StompProjectile>(damage: Damage * 2);
       }
       if (abilities.airJump.Active) {
@@ -83,11 +83,11 @@ namespace OriMod.Abilities {
       player.runSlowdown = 8;
       player.gravity = Gravity;
       player.maxFallSpeed = MaxFallSpeed;
-      oPlayer.immuneTimer = 12;
+      abilities.oPlayer.immuneTimer = 12;
     }
 
     internal void EndStomp() {
-      oPlayer.PlaySound("Ori/Stomp/seinStompImpact" + _randEnd.NextNoRepeat(3), 0.9f);
+      abilities.oPlayer.PlaySound("Ori/Stomp/seinStompImpact" + _randEnd.NextNoRepeat(3), 0.9f);
       abilities.airJump.currentCount = 0;
       player.velocity = Vector2.Zero;
       Vector2 position = new Vector2(player.position.X, player.position.Y + 32);
@@ -97,12 +97,12 @@ namespace OriMod.Abilities {
         dust.velocity *= new Vector2(6, 1.5f);
         dust.velocity.Y = -Math.Abs(dust.velocity.Y);
       }
-      PutOnCooldown();
+      StartCooldown();
       NewAbilityProjectile<StompEnd>(damage: Damage);
-      SetState(State.Inactive);
+      SetState(AbilityState.Inactive);
     }
 
-    protected override void UpdateUsing() {
+    public override void UpdateUsing() {
       player.controlUp = false;
       player.controlDown = false;
       if (Starting) {
@@ -114,19 +114,19 @@ namespace OriMod.Abilities {
       player.controlThrow = false;
       player.controlUseItem = false;
       player.controlUseTile = false;
-      oPlayer.KillGrapples();
+      abilities.oPlayer.KillGrapples();
     }
 
-    internal override void Tick() {
+    public override void PreUpdate() {
       if (Inactive) {
         if (CanUse) {
-          if (input.stomp.JustPressed) {
+          if (abilities.oPlayer.input.stomp.JustPressed) {
             _currentHoldDown = 1;
           }
-          if (_currentHoldDown >= 1 && input.stomp.Current) {
+          if (_currentHoldDown >= 1 && abilities.oPlayer.input.stomp.Current) {
             _currentHoldDown++;
             if (_currentHoldDown > HoldDownDelay) {
-              SetState(State.Starting);
+              SetState(AbilityState.Starting);
               _currentHoldDown = 0;
             }
           }
@@ -134,20 +134,17 @@ namespace OriMod.Abilities {
         if (Starting) {
           _currentHoldDown = 0;
         }
-        else {
-          TickCooldown();
-        }
       }
       else if (Starting) {
-        if (CurrentTime > StartDuration) {
-          SetState(State.Active);
+        if (stateTime > StartDuration) {
+          SetState(AbilityState.Active);
         }
       }
       else if (Active) {
-        if (CurrentTime > MinDuration && !player.controlDown || abilities.airJump) {
-          SetState(State.Inactive);
+        if (stateTime > MinDuration && !player.controlDown || abilities.airJump) {
+          SetState(AbilityState.Inactive);
         }
-        if (oPlayer.IsGrounded) {
+        if (abilities.oPlayer.IsGrounded) {
           EndStomp();
         }
       }