+++ Projectiles/Abilities/ChargeJumpProjectile.cs
@@ -7,11 +7,11 @@ namespace OriMod.Projectiles.Abilities {
   /// <summary>
   /// Projectile hitbox for when the player is using <see cref="ChargeJump"/> and <see cref="WallChargeJump"/>.
   /// </summary>
-  public sealed class ChargeJumpProjectile : AbilityProjectile {
+  public sealed class ChargeJumpProjectile : OriAbilityProjectile {
     /// <summary>
-    /// This is not used, as this <see cref="AbilityProjectile"/> is used by both <see cref="ChargeJump"/> and <see cref="WallChargeJump"/>.
+    /// This is not used, as this <see cref="OriAbilityProjectile"/> is used by both <see cref="ChargeJump"/> and <see cref="WallChargeJump"/>.
     /// </summary>
-    public override byte Id => AbilityId.ChargeJump;
+    public override int Id => AbilityId.ChargeJump;
 
     public override void SetDefaults() {
       base.SetDefaults();
@@ -20,7 +20,7 @@ namespace OriMod.Projectiles.Abilities {
     }
 
     protected override void CheckAbilityActive() {
-      if (oPlayer.abilities.chargeJump || oPlayer.abilities.wallChargeJump) {
+      if (abilities.chargeJump || abilities.wallChargeJump) {
         Projectile.timeLeft = 2;
       }
     }
@@ -28,7 +28,7 @@ namespace OriMod.Projectiles.Abilities {
     protected override void Behavior() {
       base.Behavior();
       // Stretch projectile size based on velocity
-      Vector2 vel = oPlayer.Player.velocity;
+      Vector2 vel = aPlayer.Player.velocity;
       Projectile.width = (int)Utils.Clamp(Math.Abs(vel.X) * 2.5f, 96, 250);
       Projectile.height = (int)Utils.Clamp(Math.Abs(vel.Y) * 2.5f, 96, 250);
     }