+++ Projectiles/Abilities/StompProjectile.cs
@@ -8,8 +8,8 @@ namespace OriMod.Projectiles.Abilities {
   /// <para>Ending stomp spawns a <see cref="StompEnd"/> projectile to deal damage.</para>
   /// </summary>
   // ReSharper disable once ClassNeverInstantiated.Global
-  public sealed class StompProjectile : AbilityProjectile {
-    public override byte Id => AbilityId.Stomp;
+  public sealed class StompProjectile : OriAbilityProjectile {
+    public override int Id => AbilityId.Stomp;
 
     public override void SetDefaults() {
       base.SetDefaults();
@@ -20,12 +20,12 @@ namespace OriMod.Projectiles.Abilities {
     protected override void Behavior() {
       base.Behavior();
       // Height is stretched based on velocity.
-      Projectile.height = Math.Max(56, (int)(oPlayer.Player.velocity.Y * 2));
+      Projectile.height = Math.Max(56, (int)(aPlayer.Player.velocity.Y * 2));
       Projectile.position.Y += 10;
     }
 
     public override void OnHitNPC(NPC target, int damage, float knockback, bool crit) {
-      Stomp stomp = oPlayer.abilities.stomp;
+      Stomp stomp = abilities.stomp;
       if (target.life > 0 && stomp.InUse) {
         stomp.EndStomp();
       }